backend/requirements.txt
# Dependências MÍNIMAS que funcionam em Python 3.13
fastapi==0.104.1
uvicorn==0.24.0

# JSON Web Token - versão compatível
PyJWT==2.8.0

# Password hashing - usando bcrypt simples
bcrypt==4.1.2

# ASGI CORS - mais simples que middleware FastAPI
starlette==0.27.0
